roughly exported api calls to a seperate file
